+++ example/infrastructure/mapper.go
@@ -1,8 +1,8 @@
 package infrastructure
 
 import (
-	"github.com/abekoh/mapc/examples/sqlc-domainmodel/domain"
-	"github.com/abekoh/mapc/examples/sqlc-domainmodel/infrastructure/sqlc"
+	"github.com/abekoh/mapc/example/domain"
+	"github.com/abekoh/mapc/example/infrastructure/sqlc"
 )
 
 // MapUserToUser maps User to User.
